Include XDG-patched version of vcsh
You can now set the environment variables
* `VCSH_GITIGNORE_D`
* `VCSH_GITATTRIBUTES_D`
to define the locations of the .gitignore and .gitattributes files for
vcsh.

They'll also get added to the repository for consistent rollouts;
although the code for doing that was also present in the unpatched
version, it only worked for manually defined files using `VCSH_GITIGNORE`
and `VCSH_GITATTRIBUTES`.