Add automatic mr repository setup hook
If you run `vcsh init $REPO_NAME`, the hook will check for
`$XDG_CONFIG_HOME/mr/available.d` to exist and then add a myrepos
configuration to a file in there as well as enabling said configuration
for mr.

The configuration will only be added if the environment variable
`GIT_VCSH_BASE` is set to inform about the defaul origin for vcsh
repositories.