Don't try to load the vim colorscheme too early
If the plugins haven't been installed successfully yet, automatically
ignore loading the plugin-supplied colorscheme.

Technically, we're just checking if gruvbox is in the `runtimepath`; so
it could theoretically be installed without all plugins being installed,
but as that is only relevant for the theme, we can live with that
distinction.