Add vcsh-bootstrap.sh to sparse checkout list
We also needed to rework (or rather remove) the test for the sparse
checkouts so that we can actually update the sparse checkout list on the
fly.

For facilitating upgrades between sparse checkout tags, we'll also be
adding the `sparse_checkout_staging.sh` script and prepare for that in the
sparse checkout list.